+++ src/main/java/com/buuz135/thaumicjei/config/ThaumicConfig.java
@@ -18,6 +18,9 @@ public class ThaumicConfig {
     @Config.Comment("Items blacklisted from the checking in the Aspect For ItemStack. Format: 'minecraft:stone'")
     public static String[] blacklistedFromAspectChecking = new String[]{};
 
+    @Config.Comment("If true the Aspect from ItemStack json will be created always on startup, if false if will be only be created if it doesn't exist and ")
+    public static boolean alwaysRecreateAspectFromItemStackFile = true;
+
     @Config.Comment("Allow to render the ItemStacks aspects in all GUI")
     public static boolean forceAspectTooltipInAllGUI = true;